Several changes having to do with resizing.
Don't override minSize property in classes that are able to resize. It is
inconsistant that only classes that are able to resize can have their
minSize less than that required by the filesystem metadata. Refine
minSize property in FS, and add a large comment explaining rationale.
Also, this makes free() slightly more realistic, although I think it
probably needs a little more care if we want to believe that it is as
precise as our tools can make it. (Apparent ext2's minimum size is actually
0, so much of this change really only effects ntfs).

Simplify _getMinSize() so that it is more friendly to call, and simplify
its caller, updateSizeInfo() a bit. updateSizeInfo() can not be simplified
very much, but it is possible to get rid of the messing with self._minSize
than it had previously and to use _minInstanceSize instead. This works
so nicely, because of the change to minSize, described above.

Changing definition of _resizable property, so that it only depends on
self._resize. Self._resizable is set when _getMinSize and _getExistingSize
are called, so that dependency is encoded in self._resizable. The only
remaining dependency is that self._resize must be available.

Get rid of _getFSInfo, _getExistingSize, _getMinSize entirely. They are only
used in updateSizeInfo, currently, and that method has peculiar and precise
requirements.

Signed-off-by: mulhern <amulhern@redhat.com>